+++ src/internet/model/tcp-scalable.cc
@@ -28,7 +28,6 @@
  */
 
 #include "tcp-scalable.h"
-#include "ns3/tcp-socket-base.h"
 #include "ns3/log.h"
 
 namespace ns3 {
@@ -133,7 +132,7 @@ TcpScalable::GetSsThresh (Ptr<const TcpSocketState> tcb,
   uint32_t segCwnd = bytesInFlight / tcb->m_segmentSize;
 
   double b = 1.0 - m_mdFactor;
-  uint32_t ssThresh = std::max (2.0, segCwnd * b);
+  uint32_t ssThresh = static_cast<uint32_t> (std::max (2.0, segCwnd * b));
 
   NS_LOG_DEBUG ("Calculated b(w) = " << b <<
                 " resulting (in segment) ssThresh=" << ssThresh);
